Rework USB binding
Support both ById and ByPort bindings - on attach we resolve them to bus+addr
so qemu doesn't have to scan (and we have more control).

The wizard stuff is mostly placeholders - most notably, there is no way for users
to specify that they want to bind a port right now - you have to manually edit the config.